Move Models functions into their own file
Move Models functions into their own file as pushing them into the Models class requires significant changes to the Explorer
